[iOS][PT][TabStrip] Fix placeholder gap
This CL fixes the placeholder gap while dragging/sliding a tab.
The gap should be equal to the width of the dragged tap, not to
the current tab width.
This change is done behind an "if else" block to avoid any side effects
related to the previous implementation.

Fixed: 1411227
Change-Id: I88b5b577891cbc8284a0d7489130caab7bad1517
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4203108
Reviewed-by: Christian Xu <christianxu@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Ewann Pellé <ewannpv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1099146}