[Telemetry] Remove dead version skew test code
This change removes version skew test code for
the telemetry API. This code is unused since the
crosapi service we check for is always available
in the Ash versions we test with. This code will
be replaced with proper version skew handling in
the future (currently in the design phase).

Bug: b/267154042
Change-Id: I105c17d9efbb6ffd707904ede5e78cac34a0a9ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4203332
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Commit-Queue: Bastian Kersting <bkersting@google.com>
Cr-Commit-Position: refs/heads/main@{#1099250}