[suggest-internals] Moves search requests to RemoteSuggestionsService
Moves the Suggest request logic for typed suggestions in SearchProvider
into the RemoteSuggestionsService. As of this change, all of the Suggest
requests with the exception of mobile Start modules happen in one place.

For the time being, zero-prefix and typed requests are done in two
nearly identical methods. This is done to avoid making non-functional
changes to the network annotations and request flags. These two methods
should be merged into a single one in a follow up CL.

One noteworthy change to the typed request is that we now always pass
InIncognito::kNo to variations::AppendVariationsHeaderUnknownSignedIn().
This is OK since suggest requests are NOT made in incognito mode.

Bug: 1410570
Change-Id: Ib67f25815bac38d5e0bdff28cba0d6dd1c3c1e12
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4202053
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1099253}