[ios][userpolicy] User policy prompt half sheet
The prompt is used to notify the user that from now on their
administrator may push policies to their browser when they use a
managed account.

The user is offered 2 options:
 - Continue
 - Sign Out and Clear Data

The User Policy feature isn't yet released. There is still UX work to
be done on the prompt, but this initial iteration is a good start.
The layout will probably remain as is and the text is subject to
change.

User Policy isn't a policy in itself where it adds a new provider to
get policies using a managed account.

Mock: http://shortn/_YimzhZuC57
Implementation: http://shortn/_QYMgdC3T59
Demo Video: http://shortn/_Csim1j4PoY

Bug: 1410811
Change-Id: Idcc1df3cafd0164309c315d70503236a0b3342de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4198119
Commit-Queue: Vincent Boisselle <vincb@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1099282}