[User Education] Add testing method to verify anchor element is present
With the addition of nested targeting for anchor elements, it has become easier to decouple the mixin-implementing component from the intended anchor element. This unfortunately makes it easier to introduce brittle anchoring that can be broken without developer knowledge. This test method is meant to be used by default everywhere that webui help bubbles are registered to ensure a test will fail if the anchor element is altered.

Bug: 1410335
Change-Id: I112fed368e4db3151a3946234060dfde5e63486f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4195031
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Mickey Burks <mickeyburks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1099307}