[NTP][Panorama] Fix header spacing
Set the header to have the same height as the line height for the header text. Otherwise, the height of the entire clickable button area (36px) increases the height of the header area, which makes the empty space above and below the header much larger.

Also, increased the top margin for the refresh daily toggle to be the total margin between itself and the header in order to account for css margin collapse. Otherwise, we end up with a total of 16px (the larger of the 2 margins) between the 2 elements, rather than the intended 24 px (16px + 8px). By setting the refresh daily top margin to 24px, the 16px bottom margin of the header will be collapsed, leaving us with a total 24px margin.

Screenshots:
screenshot/4q5RHPRqLS55SiF
screenshot/p8appAYKUrtxkV6
screenshot/9dBUa6ZMoHwupxp
screenshot/7HH4Qz82VniBhDT

Bug: 1406563
Change-Id: I2990219a640aad8bbe11ad338532e88f3e873589
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4206658
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Commit-Queue: Riley Tatum <rtatum@google.com>
Cr-Commit-Position: refs/heads/main@{#1099308}