Add test for cancel from start of printing
Add test coverage for case where a cancel occurs right at the start of
printing.  This can happen at least for Windows, for printer drivers
which print to file and thus display a dialog to choose the filename.
The user is able to cancel out at this point.

Illustrates that an error dialog is shown when this situation occurs,
even though the cancel was due to a user action.

Bug: 1393505
Change-Id: Iaf751373f650f5cdc5a6ead67582d64766687b4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4205936
Reviewed-by: Andy Phan <andyphan@chromium.org>
Commit-Queue: Alan Screen <awscreen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1099341}