[iOS] Fixed Price Tracking IPH Not Displaying
The Price Tracking IPH was not displaying when the user navigated to a
price trackable product page for the first time. The reason for this was
because the IPH's trigger configuration's availability property required
that the IPH be available for at least seven days. This essentially
delayed the Price Tracking's first IPH by N + 7 days.

In this CL, the IPH's availability property was set to ANY and a new
`EventConfig` was added that replaces the availability property's
original purpose of ensuring that there is a 7 day gap between
impressions.

Fixed: 1411715
Change-Id: I8d56b6b65e7ad79ac9ab7c8d3ff48a6d45a486c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4209188
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Daniel White <danieltwhite@google.com>
Cr-Commit-Position: refs/heads/main@{#1099376}