multipaste: Cache main and display formats for clipboard history items
The function that calculates a ClipboardData's main format for use in
clipboard history returns an optional. Currently, most call sites do not
check whether that optional has a value, which has not caused a problem
because items are guaranteed to have a data format in production code.

This CL updates ClipboardHistoryItem to calculate, store, and return its
own main and display formats, eliminating the need for optional checking
in subsequent format queries on the item. This CL also updates the API
and uses of the ClipboardHistoryItemBuilder test utility to accommodate
the new invariant that all items have at least one data format.

Bug: b/266947643
Change-Id: I472a5aea4569ace41a795e6378adf24e3ce23e91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4206648
Reviewed-by: David Black <dmblack@google.com>
Commit-Queue: Colin Kincaid <ckincaid@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1099381}