Fenced frames: Update fenced-frame-mparch web platform tests to test `selectURL()` that returns a FencedFrameConfig.
crrev.com/c/4111682 enables selectURL to return a FencedFrameConfig
object by specifying the new boolean dictionary member in
SharedStorageRunOperationMethodOptions as true.

This CL updates the existing tests that directly call generateURN().
The implementation of generateURN() is to run a mock selectURL.
It is updated to test both the existing behavior (returning an
urn:uuid) and new behavior (returning a fenced frame config).

generateURN() is also renamed to runSelectURL() to inform the client
that a selectURL() will run when calling this function. As there will
be a limit of 3 calls on selectURL() enforced in the future.

Please note:
1. Test urn is renamed to loading, as this test ultimately tests
   the loading of a fenced frame. Now we can not only use urn:uuid,
   but also fenced frame config to do this.
2. Test csp-urn is split into two files to prevent time-out:
  a. csp-allowed
  b. csp-blocked
3. Add `once: true` to make the csp event listener one-off.
4. Moved some common functions to utils.js.
5. Renamed generateURN to runSelectURL and updated its implementation.
6. Move tests that use selectURL in disallowed-navigations to
   separate files.
  a. disallowed-navigation-to-blob
  b. disallowed-navigation-to-data
  c. This will also make future changes easier when the limit of
     number of selectURL() calls are enforced.

Change-Id: I93e343c54a563fca3f114844cd6dec0a7f5d21b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4197923
Commit-Queue: Xiaochen Zhou <xiaochenzh@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1099389}