Refactor WebAppFrameToolbarView ownership 5/x: Windows.
And "finally", this also makes PWAs work on Windows with the
WebAppFrameToolbarInBrowserView feature enabled. Not quite finally,
since tests still need to be fixed in more CLs to work with the
flag enabled.

Also includes a minor cleanup in the Mac code, removing a redundant
check (GetIsWebAppType is exactly the same check as checking if the
browser has an app controller).

Bug: 1407240
Change-Id: Ib7ce9d1d55533056c5dc7c2ae7792063b8a890a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4193138
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1099427}