[Extensions] Re-enable some TabDuplicate tests.
These tests have been disabled for a long time due to flakiness. The
fix is to wait for tabs to be loaded before checking their properties.

One test is still disabled because it causes the browser to DCHECK.
It looks like a change has broken duplicating a tab from a popup
window. This will require more investigation as to the motivation
for the change that broke this behavior.

Bug: 149924
Change-Id: I94c990aa1ea893e35faceece7c89ebac482a3460
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4200834
Reviewed-by: Tim <tjudkins@chromium.org>
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1099440}