Add StyledLabel::SetDisplayedOnBackgroundColorId.
This CL adds a sister method to the existing
SetDisplayedOnBackgroundColor that uses color ids instead of colors with
the goal of eliminating custom OnThemeChanged methods in classes that
use StyledLabel.

It also fixes a bug in label in which SetBackgroundColorId does not
properly update the colors of the label and adds unit tests.

Bug: 1411172
Change-Id: I3e0c83bf15f4a6d85ffde142cbee6a30c25428a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4200658
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1099458}