Fixed keyboard mute issue with volume slider
This CL fixed the issue that the keyboard mute key won't make the volume
sliders(in the main page, in the toast, in the audio subpage) visually
muted. The expected behavior is: if press the mute key, the slider will
go to 0 point and the icon becomes muted, then if press the volume up
key, the slider will restore its previous state. If press the volume
down key after mute, the slider will preserve the mute state. Also added
some unittest for these behaviors.

Bug: b/266718867
Change-Id: Icc99dc28347ee593d84e1c19490d17b78fe0da16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4206145
Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Sylvie Liu <sylvieliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1099482}