Floss: Rename GattClient to GattManagerClient
Before, the naming convention of GattClient would cause confusion
between "a client for managing Gatt" and "a Gatt Client that can
interact with a Gatt Server".

With this change, GattClient is renamed to GattManagerClient, making
the logical difference between the two concepts apparent in the
source code.

BUG=b:193686557

Test: Successful compilation of Floss unit tests
Change-Id: I55318199a8dcf70dacaec6608b36087c5625e535
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4206987
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>
Commit-Queue: Sarvesh Kalwit <sarveshkalwit@google.com>
Cr-Commit-Position: refs/heads/main@{#1099489}