virtual_desks: Update save desk buttons on starting animations complete
In some use cases, the desks bar is initialized immediately when
entering overview instead of on the starting animations complete. For
example, the active desk has a fullscreen window. If we update save
desk buttons when initializing the desks bar, the origin bounds of the
button could be incorrect, since by the time the target bounds of the
first overview item could be not set yet.

To fix this, update the save desk buttons on starting animations
complete instead of while initializing the desks bar.

Bug: b/267234552
Change-Id: Iea6691281a81f45426bef6f8fb4be0dbfe2de412
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4209856
Commit-Queue: Connie Xu <conniekxu@chromium.org>
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1099498}