[Phone Hub][Eche] Track when glimmer animation or apps button is shown
This histogram tracks when the user opens the Phone Hub bubble whether
or not the more apps button is showing the glimmer animation or the
loaded app state.

When opening the bubble and it shows the animation but quickly
transitions to the loaded app state view, it will log as the animation
showing correctly.

Flagged: kEcheLauncherIconsInMoreAppsButton
Test: manually verified both states on chrome://histograms
Change-Id: I0e0d75beef63f2d55909123ec4a07efbaf875f7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4191050
Reviewed-by: Jon Mann <jonmann@chromium.org>
Commit-Queue: Crisrael Lucero <crisrael@google.com>
Cr-Commit-Position: refs/heads/main@{#1099540}