[omnibox] Change button row view buttons from one pedal to many actions
This CL reworks OmniboxSuggestionButtonRowView to use a vector of
action buttons instead of one `pedal_button`. Actions from the
match are associated, and the view is only rebuilt when there is
a structural change. This is more flexible and more efficient
than the previous implementation because most matches have zero
actions and don't need to allocate a button row with action views,
but when there are actions, there may be any number, not just one.

Bug: 1408506
Change-Id: I87954ef2360dd550e285eb5b7cfa6794d3df187a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4201287
Reviewed-by: Angela Yoeurng <yoangela@chromium.org>
Commit-Queue: Orin Jaworski <orinj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1099601}