Notify Files app when default tasks are changed
Implement the functionality added in https://crrev.com/c/3698346 for
when the default task is changed for a particular mime type or file
suffix.

This solves the bug that occurs when the default task for an office file
is set by the Office Setup flow but the Files app isn't made aware.

Bug: b:266667405
Test: Manual Test
Change-Id: Id87e063ff9e0ce1b7b7b03579aca3caa90273176
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4207468
Commit-Queue: Cassy Chun-Crogan <cassycc@google.com>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1099632}