[omnibox][grouping] Replace `AdjustLimitsAndResetCounts()` flow.
We need a mechanism to prioritize some groups over others. E.g., if we
have a default suggestion, we want to make sure we show it even if there
are other search or history suggestions with higher relevance.

For non-zps, priority is based on their relevance, allowed to be
default, and match type (for page classification based demotion in
realbox).

We enforce this priority by sorting matches according to their priority.
E.g., the default suggestion is ordered above the remaining suggestions.

For zps, priority additionally factors in match group. This isn't
accounted for in the pre-sort. Instead, the
`AdjustLimitsAndResetCounts()` flow ensures the correct zps priority by
adjusting each groups limits to the min of how many assignable matches
it has and the remaining section capacity after allocating the limits of
higher priority groups. The implementation is a bit complicated. It
requires pseudo-adding matches to sections to predict the count for each
group, then calcuating the groups limits, and lastly resetting the
groups in preparation for the 'real' pass.

This CL instead uses the non-zps approach of ordering matches according
to their groups' priorities.

Bug: 1343512
Change-Id: I485f43070a8f61543b0564488a9558ea0c3fab39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4201269
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1099667}