Record CacheTransparency UKMs in browser process
This CL passes the UKM information related to Cache Transparency from
the renderer to the browser process, so that they are logged in the same
process as where the rest of the Page Load Metric UKMs are recorded,
to avoid running into downsampling issues when analyzing metrics. To
do this, SubresourceLoadMetrics is used to store the required
information.

We also add this information to DeferredFinishLoadingInfo in
ResourceLoader to make sure we can pass the correct information to
the browser process after all the body is received.

In this CL, we:
- deprecate the old Network.CacheTransparency UKM event
- add a new Network.CacheTransparency2 UKM event
- add a new page load metrics observer (PLMO) for Cache Transparency

UKM collection review doc: https://docs.google.com/document/d/1ZhhDiqvSwab6SG32tEItw9CRYQBjuAGDBOCGk0MPTAU/edit?usp=sharing&resourcekey=0-Z7t00UitAV17yuCEOKDR9w

Bug: 1309002
Change-Id: Ia98cce5d7341898e923fb4e6f2a0c5643c0a63d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4161203
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Nidhi Jaju <nidhijaju@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1099680}