[GPU] Have command decoders call BindTexImage() only for GLImageD3D
The command decoders do lazy binding only on Windows/Mac. On these
platforms, the only GLImage subclass with a non-empty implementation of
BindTexImage() is GLImageD3D. This CL restricts the decoders to calling
BindTexImage() *only* if the GLImage is a GLImageD3D instance as a step
toward eliminating BindTexImage() from the GLImage interface.

Note: I believe that in practice a GLImageD3D instance is never set as
the level image with a bind pending, but I leave removal of these
callsites to be done as part of crbug.com/1410157.

Bug: 1310018
Change-Id: I404938ebd2dda501981a2c95520def10ebe963b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4206432
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1099688}