Restructure VFHPL::OnFrameReadyInBuffer to match other functions
Though the crash appears to not be reproing anymore, it appears to be
due to an AV on the handler_ member. OnFrameReadyInBuffer appears to be
the only method that only checked first handler_in_process_ before
returning and then always running the handler_ path if not. Every other
function was structured with an if/elseif rather than an if/return.
Given the CHECK in the constructor, the if/return pattern *should* have
been fine; but it's possible that there is a shutdown race or that there
was otherwise some kind of lacros version mismatch.

This patches over the most likely cause of the crash by converting the
if (handler_in_process_) { ... return; } pattern to match the other
functions in this file by following a if (handler_.is_bound() {} else
if (handler_in_process_) {} pattern. Note that because of the ivnersion
of the order of the if statements and the additional nesting level, the
diff looks larger than it realistically should.

Fixed: 1383999
Change-Id: Id23f1106b6d305aae230b32c2c0473645fe754f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4210868
Reviewed-by: Markus Handell <handellm@google.com>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Markus Handell <handellm@google.com>
Cr-Commit-Position: refs/heads/main@{#1099701}